requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
amqp==5.3.1
attrs==24.3.0
bcrypt==4.3.0
bidict==0.23.1
billiard==4.2.1
blinker==1.8.2
cachelib==0.13.0
cachetools==5.5.0
celery==5.5.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==44.0.2
Deprecated==1.2.14
Faker==37.5.3
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
frozenlist==1.5.0
geoip2==4.8.1
gevent==24.11.1
google-api-core==2.23.0
google-auth==2.36.0
google-cloud==0.34.0
google-cloud-recaptcha-enterprise==1.24.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.5.4
libmagic==1.0
limits==3.13.0
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
maxminddb==2.8.2
mdurl==0.1.2
msgspec==0.19.0
multidict==6.1.0
mysql-connector-python==9.1.0
ordered-set==4.1.0
packaging==24.1
paramiko==3.5.1
pillow==11.0.0
pluggy==1.5.0
prompt_toolkit==3.0.51
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.10.1
PyMySQL==1.1.1
PyNaCl==1.5.0
pyotp==2.9.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-magic==0.4.27
python-socketio==5.12.1
qrcode==8.0
redis==5.2.0
requests==2.32.3
rich==13.9.4
rsa==4.9
sentry-sdk==2.18.0
setuptools==75.5.0
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.36
stripe==11.2.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
urllib3==2.2.3
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.0
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.2.1
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
